New attack, bunch of improvements.

+++ src/controllers/accountController.py
@@ -12,6 +12,7 @@ from werkzeug.exceptions import Forbidden, BadRequest
 from werkzeug.utils import secure_filename
 import uuid
 import os
+import mimetypes
 
 
 
@@ -127,7 +128,28 @@ def orders():
 
 
 #MaliciousFileUpload-1 - START  
-"""Vulnerability"""
+"""Fix"""
+ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif', 'svg'}
+ALLOWED_MIME = {'image/png','image/jpg','image/jpeg', 'image/svg+xml'}
+
+def check_extension(filename):
+    """Function returns True if extension is all right, and False if otherwise."""
+    filename_split = filename.split('.')
+    if len(filename_split) != 2:
+        return False
+    else:
+        extension = filename_split[-1]
+        if extension in ALLOWED_EXTENSIONS:
+            return True
+        else:
+            return False
+
+def check_mime(picture):
+    if mimetypes.guess_type(picture.filename)[0] in ALLOWED_MIME:
+        return True
+    else:
+        return False
+
 def upload_picture():
     if request.method == 'POST':
         try:
@@ -137,6 +159,7 @@ def upload_picture():
             # check if the post request has the file part
             if 'profile_picture' not in request.files:
                 flash('No file part', 'danger')
+                print('No file part')
                 return redirect(request.referrer)
             picture = request.files['profile_picture']
             #print("Filename: ", picture.filename)
@@ -144,26 +167,32 @@ def upload_picture():
             # empty file without a filename.
             if picture.filename == '':
                 flash('No selected file', 'danger')
+                print('No selected file')
                 return redirect(request.referrer)
-            picture_uuid = str(uuid.uuid1())+ picture.filename
-            if user.profile_picture:
+            filename_sanitized = secure_filename(picture.filename)
+            print("Sanitized: ", filename_sanitized)
+            if picture and check_extension(filename_sanitized) and check_mime(picture):
+                file_to_store = str(uuid.uuid1())+filename_sanitized
+                if user.profile_picture:
                     original_picture = user.profile_picture
                     print(original_picture)
                     path_to_original_picture = os.path.join(UPLOAD_FOLDER,original_picture)
                     print(path_to_original_picture)
                     os.remove(path_to_original_picture)
-            user.profile_picture = picture_uuid
-            db.session.commit()
-            upload_path = os.path.join(UPLOAD_FOLDER,picture_uuid)
-            picture.save(upload_path)
-            flash('Profile picture has been updated.', 'success')
-            log_config.logger.info("User %s successfully updated his profile picture." % user.username, extra={'ip_address': request.remote_addr})
-            return redirect(request.referrer)
+                user.profile_picture = file_to_store
+                db.session.commit()
+                upload_path = os.path.join(UPLOAD_FOLDER,file_to_store)
+                picture.save(upload_path)
+                flash('Profile picture has been updated.', 'success')
+                log_config.logger.info("User %s successfully updated his profile picture." % user.username, extra={'ip_address': request.remote_addr})
+                return redirect(request.referrer)
+            else:
+                return BadRequest()
         except ValidationError:
             log_config.logger.error("User was not updated. Missing or invalid CSRF token.", extra={'ip_address': request.remote_addr})
             return Forbidden()
         except Exception as e:
-            flash("Error occureed. Please try again.", 'danger')
+            flash("Error occureed. Please try again.",'danger')
             log_config.logger.error("User was not successfully updated. Exception: %s" % e, extra={'ip_address': request.remote_addr})
             return redirect(request.referrer) 
 #MaliciousFileUpload-1 - END
